chore: Remove unused functions.
