+++ Assets/UnZENity-Core/Scripts/Caches/TextureCache.cs
@@ -1,11 +1,7 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using GUZ.Core.Data.Container;
 using GUZ.Core.Extensions;
 using GUZ.Core.Manager;
 using GUZ.Core.Util;
@@ -14,7 +10,6 @@ using UnityEngine;
 using UnityEngine.Rendering;
 using ZenKit;
 using Debug = UnityEngine.Debug;
-using Mesh = UnityEngine.Mesh;
 using Object = UnityEngine.Object;
 using Texture = UnityEngine.Texture;
 using TextureFormat = UnityEngine.TextureFormat;
@@ -48,8 +43,6 @@ namespace GUZ.Core.Caches
         public const int MaxMipCount = 6;
 
         public static Dictionary<TextureArrayTypes, Texture> TextureArrays { get; } = new();
-        public static List<(MeshRenderer Renderer, WorldContainer.SubMeshData SubmeshData)> WorldMeshRenderersForTextureArray = new();
-        public static Dictionary<Mesh, VobMeshData> VobMeshesForTextureArray = new();
 
         private static readonly Dictionary<string, Texture2D> _texture2DCache = new();
         private static readonly Dictionary<TextureArrayTypes, List<(string PreparedKey, ZkTextureData TextureData)>> _texturesToIncludeInArray = new();
@@ -226,136 +219,6 @@ namespace GUZ.Core.Caches
             }
         }
 
-        public static async Task BuildTextureArrays()
-        {
-            // Unload the previous texture arrays.
-            foreach (TextureArrayTypes item in TextureArrays.Keys)
-            {
-                Object.Destroy(TextureArrays[item]);
-            }
-            TextureArrays.Clear();
-
-            Stopwatch stopwatch = new();
-            stopwatch.Start();
-
-            foreach (TextureArrayTypes textureArrayType in _texturesToIncludeInArray.Keys)
-            {
-                if (Application.isEditor)
-                {
-                    for (int i = 16; i <= 2048; i *= 2)
-                    {
-                        int resCount = _texturesToIncludeInArray[textureArrayType].Where(t => Mathf.Max(t.TextureData.Width, t.TextureData.Height) == i).Count();
-                        Debug.Log($"[{nameof(TextureCache)}] {textureArrayType}: {resCount} textures ({Mathf.RoundToInt(100 * (float)resCount / _texturesToIncludeInArray[textureArrayType].Count)}%) with dimension {i}.");
-                    }
-                }
-
-                // Create the texture array with the max size of the contained textures, limited by the max texture size.
-                int maxSize = Mathf.Min(MaxTextureSize, _texturesToIncludeInArray[textureArrayType].Max(p => p.TextureData.Width));
-                ZkTextureData textureWithMaxAllowedSize = _texturesToIncludeInArray[textureArrayType].Where(t => t.TextureData.Width == maxSize && t.TextureData.Height == maxSize).Select(t => t.TextureData).FirstOrDefault();
-                // Find the max mip depth defined for the max allowed texture size.
-                int maxMipDepth = 0;
-                if (textureWithMaxAllowedSize == null)
-                {
-                    Debug.LogError($"[{nameof(TextureCache)}] No texture with size {maxSize}x{maxSize} was found to determine max allowed mip level. Falling back to texture with highest mip count.");
-                    maxMipDepth = _texturesToIncludeInArray[textureArrayType].Max(t => t.TextureData.MipmapCount);
-                }
-                else
-                {
-                    maxMipDepth = textureWithMaxAllowedSize.MipmapCount;
-                }
-
-                TextureFormat textureFormat = TextureFormat.RGBA32;
-                if (textureArrayType == TextureArrayTypes.Opaque)
-                {
-                    textureFormat = TextureFormat.DXT1;
-                }
-
-                Texture texArray;
-                if (textureArrayType != TextureArrayTypes.Water)
-                {
-                    texArray = new Texture2DArray(maxSize, maxSize, _texturesToIncludeInArray[textureArrayType].Count, textureFormat, maxMipDepth, false, true)
-                    {
-                        filterMode = FilterMode.Trilinear,
-                        wrapMode = TextureWrapMode.Repeat
-                    };
-                }
-                else
-                {
-                    texArray = new RenderTexture(maxSize, maxSize, 0, RenderTextureFormat.ARGB32, maxMipDepth)
-                    {
-                        dimension = TextureDimension.Tex2DArray,
-                        autoGenerateMips = false,
-                        filterMode = FilterMode.Trilinear,
-                        useMipMap = true,
-                        volumeDepth = _texturesToIncludeInArray[textureArrayType].Count,
-                        wrapMode = TextureWrapMode.Repeat
-                    };
-                }
-
-                // Copy all the textures and their mips into the array. Textures which are smaller are tiled so bilinear sampling isn't broken - this is why it's not possible to pack different textures together in the same slice.
-                for (int i = 0; i < _texturesToIncludeInArray[textureArrayType].Count; i++)
-                {
-                    Texture2D sourceTex = TryGetTexture(_texturesToIncludeInArray[textureArrayType][i].PreparedKey, false);
-
-                    int sourceMip = 0;
-                    int sourceMaxDim = Mathf.Max(sourceTex.width, sourceTex.height);
-                    int sourceWidth = sourceTex.width;
-                    int sourceHeight = sourceTex.height;
-                    while (sourceMaxDim > MaxTextureSize)
-                    {
-                        sourceMip++;
-                        sourceMaxDim /= 2;
-                        sourceWidth /= 2;
-                        sourceHeight /= 2;
-                    }
-
-                    for (int mip = sourceMip; mip < sourceTex.mipmapCount; mip++)
-                    {
-                        int targetMip = mip - sourceMip;
-                        for (int x = 0; x < texArray.width / sourceWidth; x++)
-                        {
-                            for (int y = 0; y < texArray.height / sourceHeight; y++)
-                            {
-                                if (texArray is Texture2DArray)
-                                {
-                                    Graphics.CopyTexture(sourceTex, 0, mip, 0, 0, sourceTex.width >> mip,
-                                        sourceTex.height >> mip, texArray, i, targetMip, (sourceTex.width >> mip) * x,
-                                        (sourceTex.height >> mip) * y);
-                                }
-                                else
-                                {
-                                    CommandBuffer cmd = CommandBufferPool.Get();
-                                    RenderTexture rt = (RenderTexture)texArray;
-                                    cmd.SetRenderTarget(new RenderTargetBinding(new RenderTargetSetup(rt.colorBuffer, rt.depthBuffer, targetMip, CubemapFace.Unknown, i)));
-                                    Vector2 scale = new Vector2((float)sourceTex.width / texArray.width, (float)sourceTex.height / texArray.height);
-                                    Blitter.BlitQuad(cmd, sourceTex, new Vector4(1, 1, 0, 0), new Vector4(scale.x, scale.y, scale.x * x, scale.y * y), mip, false);
-                                    Graphics.ExecuteCommandBuffer(cmd);
-                                    cmd.Clear();
-                                    CommandBufferPool.Release(cmd);
-                                }
-                            }
-                        }
-                    }
-
-                    Object.Destroy(sourceTex);
-
-                    await FrameSkipper.TrySkipToNextFrame();
-                }
-
-                TextureArrays.Add(textureArrayType, texArray);
-            }
-
-            // Clear cached texture data to save storage.
-            foreach (List<(string PreparedKey, ZkTextureData Texture)> textureList in _texturesToIncludeInArray.Values)
-            {
-                textureList.Clear();
-                textureList.TrimExcess();
-            }
-
-            stopwatch.Stop();
-            Debug.Log($"Built tex array in {stopwatch.ElapsedMilliseconds / 1000f} s");
-        }
-
         public static async Task BuildTextureArray()
         {
             await BuildTextureArray(TextureFormat.DXT1,
@@ -493,52 +356,6 @@ namespace GUZ.Core.Caches
             return TextureArrays[textureArrayType];
         }
 
-        private static bool GetAlreadyIncludedTexture(TextureArrayTypes textureArrayType, string key, out int arrayIndex, out int maxMipLevel, out Vector2 textureScale, out int animFrameCount)
-        {
-            if (_texturesToIncludeInArray.ContainsKey(textureArrayType))
-            {
-                (string, ZkTextureData) cachedTextureData = _texturesToIncludeInArray[textureArrayType].FirstOrDefault(k => k.PreparedKey == key);
-                if (cachedTextureData != default)
-                {
-                    arrayIndex = _texturesToIncludeInArray[textureArrayType].IndexOf(cachedTextureData);
-                    ZkTextureData zkData = _texturesToIncludeInArray[textureArrayType][arrayIndex].TextureData;
-                    maxMipLevel = zkData.MipmapCount - 1;
-                    textureScale = zkData.Scale;
-                    animFrameCount = zkData.AnimFrameCount;
-                    return true;
-                }
-            }
-
-            arrayIndex = -1;
-            maxMipLevel = -1;
-            textureScale = default;
-            animFrameCount = 0;
-            return false;
-        }
-
-        private static List<(string, ZkTextureData)> TryGetAnimTextures(string key)
-        {
-            List<(string, ZkTextureData)> textures = new();
-            if (!key.Contains("_A0") && !key.Contains("_a0"))
-            {
-                return textures;
-            }
-            for (int id = 1; ; id++)
-            {
-                // Replace the frame number in the key with the current id
-                string frameKey = Regex.Replace(key, "_[Aa]0", $"_A{id}");
-                frameKey = frameKey.ToUpper();
-                ITexture zkTex = ResourceLoader.TryGetTexture(frameKey);
-                if (zkTex == null)
-                {
-                    break;
-                }
-                textures.Add((frameKey, new ZkTextureData(frameKey, zkTex)));
-            }
-
-            return textures;
-        }
-
         /// <summary>
         /// Unity doesn't want to create mips for DXT1 textures. Recreate them as RGB24.
         /// </summary>
@@ -554,18 +371,6 @@ namespace GUZ.Core.Caches
             Object.Destroy(dxtTexture);
 
             return texture;
-        }
-
-        /// <summary>
-        /// Once a TextureArray is build and assigned to renderers, we can safely clear these lists to free managed memory.
-        /// </summary>
-        public static void RemoveCachedTextureArrayData()
-        {
-            _texturesToIncludeInArray.Clear();
-            _texturesToIncludeInArray.TrimExcess();
-
-            WorldMeshRenderersForTextureArray.Clear();
-            WorldMeshRenderersForTextureArray.TrimExcess();
         }
 
         private static string GetPreparedKey(string key)
