--- Assets/UnZENity-Core/Scripts/Creator/Meshes/Builder/WorldMeshBuilder2.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 9f7664ecb3de4c6dbd8dbad8552fbb69
-timeCreated: 1735712658
